+++ src/Enums/Host.php
@@ -3,17 +3,17 @@
 namespace Timirey\XApi\Enums;
 
 /**
- * Enum representing the WebSocket host URLs for xStation5 API.
+ * Enum representing the socket host URLs for xStation5 API.
  */
 enum Host: string
 {
     /**
      * Demo account host URL.
      */
-    case DEMO = 'wss://ws.xtb.com/demo';
+    case DEMO = 'ssl://xapi.xtb.com:5124';
 
     /**
      * Real account host URL.
      */
-    case REAL = 'wss://ws.xtb.com/real';
+    case REAL = 'ssl://xapi.xtb.com:5112';
 }