+++ src/Exceptions/SocketException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Timirey\XApi\Exceptions;
+
+use Exception;
+
+/**
+ * Custom exception for sockets.
+ */
+class SocketException extends Exception
+{
+}